Delete unused modals and associated folders. Move remaining files into parents folder.
