+++ src/pages/caregiver.tsx
@@ -6,9 +6,9 @@ import Link from "next/link";
 import { useState } from "react";
 import Footer from "@/components/layout/footer";
 import SearchEngine from "@/components/engines/searchEngine";
-import FindTab from "@/components/caregiver/tabs/findTab";
-import HistoryTab from "@/components/caregiver/tabs/historyTab";
-import ActiveTab from "@/components/caregiver/tabs/activeTab";
+import FindTab from "@/components/caregiver/findTab";
+import HistoryTab from "@/components/caregiver/historyTab";
+import ActiveTab from "@/components/caregiver/activeTab";
 
 const Caregiver: NextPage = (props) => {
   const { data: session } = useSession();
